Swap app to Flask Python App
