+++ 3-Docker/Dockerfile
@@ -1,14 +1,12 @@
-FROM python:latest
+FROM python:3.9-slim
 
 RUN mkdir /build
 WORKDIR /build
 
 COPY app /build
 
-COPY app/requirements.txt /build
-
-RUN pip install --trusted-host pypi.org --trusted-host pypi.python.org --trusted-host files.pythonhosted.org -r requirements.txt
+RUN pip install --no-cache-dir -r requirements.txt
 
 EXPOSE 5000
 
-CMD [ "python", "app.py" ]
+CMD ["python", "app.py"]
