--- kubernetes_manifest/deployment.yml
@@ -1,42 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: uber-ui
----
-
-apiVersion: v1
-kind: Service
-metadata:
-  name: uber-ui
-  namespace: uber-ui
-spec:
-  selector:
-    app: uber-ui
-  ports:
-  - protocol: "TCP"
-    port: 80
-    targetPort: 5000
-  type: LoadBalancer
-
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: uber-ui
-  namespace: uber-ui
-spec:
-  selector:
-    matchLabels:
-      app: uber-ui
-  replicas: 2
-  template:
-    metadata:
-      labels:
-        app: uber-ui
-    spec:
-      containers:
-      - name: uber-ui
-        image: devopsthehardwaytamopsacr.azurecr.io/uberapp:v1
-        ports:
-        - containerPort: 5000
----
